requirements.txt
onnx
onnxruntime==1.15.1
onnxruntime-gpu
opencv-python==4.8.0.76
numpy==1.24.3
scikit-image
gspread 
google-auth
google-auth-oauthlib
python-telegram-bot==13.7
openpyxl
flask==2.3.2
flask-socketio==5.3.4
pandas==2.0.3
aiohttp==3.8.5
